Stripped all reference to Win32, compiles on Linux, doesn't run very far.

+++ Lawn/System/ReanimationLawn.cpp
@@ -99,7 +99,7 @@ MemoryImage* ReanimatorCache::MakeBlankMemoryImage(int theWidth, int theHeight)
 	MemoryImage* aImage = new MemoryImage();
 
 	int aBitsCount = theWidth * theHeight;
-	aImage->mBits = new unsigned long[aBitsCount + 1];
+	aImage->mBits = new uint32_t[aBitsCount + 1];
 	aImage->mWidth = theWidth;
 	aImage->mHeight = theHeight;
 	aImage->mHasTrans = true;
