+++ Sexy.TodLib/FilterEffect.cpp
@@ -91,7 +91,7 @@ void FilterEffectDisposeForApp()
 //0x446FD0
 void FilterEffectDoLumSat(MemoryImage* theImage, float theLum, float theSat)
 {
-	unsigned long* ptr = theImage->mBits;
+	uint32_t* ptr = theImage->mBits;
 	for (int y = 0; y < theImage->mHeight; y++)
 	{
 		for (int x = 0; x < theImage->mWidth; x++)
@@ -126,7 +126,7 @@ void FilterEffectDoLessWashedOut(MemoryImage* theImage)
 //0x447190
 void FilterEffectDoWhite(MemoryImage* theImage)
 {
-	unsigned long* ptr = theImage->mBits;
+	uint32_t* ptr = theImage->mBits;
 	for (int y = 0; y < theImage->mHeight; y++)
 		for (int x = 0; x < theImage->mWidth; x++)
 			*ptr++ |= 0x00FFFFFF;
@@ -139,7 +139,7 @@ MemoryImage* FilterEffectCreateImage(Image* theImage, FilterEffect theFilterEffe
 	aImage->mWidth = theImage->mWidth;
 	aImage->mHeight = theImage->mHeight;
 	int aNumBits = theImage->mWidth * theImage->mHeight;
-	aImage->mBits = new unsigned long[aNumBits + 1];
+	aImage->mBits = new uint32_t[aNumBits + 1];
 	aImage->mHasTrans = true;
 	aImage->mHasAlpha = true;
 	memset(aImage->mBits, 0, aNumBits * 4);
