+++ SexyAppFramework/imagelib/ImageLib.h
@@ -1,6 +1,7 @@
 #ifndef __IMAGELIB_H__
 #define __IMAGELIB_H__
 
+#include <cstdint>
 #include <string>
 
 namespace ImageLib
@@ -11,7 +12,7 @@ class Image
 public:
 	int						mWidth;
 	int						mHeight;
-	unsigned long*			mBits;
+	uint32_t*				mBits;
 
 public:
 	Image();
@@ -19,7 +20,7 @@ public:
 
 	int						GetWidth();
 	int						GetHeight();
-	unsigned long*			GetBits();
+	uint32_t*				GetBits();
 };
 
 bool WriteJPEGImage(const std::string& theFileName, Image* theImage);
